Create PastaProb.java
The problem: Johnny goes to his local dining court 3 out of every 4 days and gets a pasta bowl. Each time, he is given a ticket numbered 1-100 along with his pasta order. Johnny is currently a sophomore and is hoping to get every ticket (i.e. at least one ticket 1, 2, and so on to 100) by the time he graduates, let's say 900 days from now. What is the expected number of days it takes to complete his task? The probability he does so successfully?

Not knowing how to do the problem, I coded this Monte Carlo simulation to get an answer. Given a sufficiently large sample size, the average number of days required to complete his goal should converge to the expected value by the Law of Large Numbers.